requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
requests==2.31.0
openai==1.32.2
PyMuPDF==1.22.2
python-docx==0.8.12
spacy==3.7.1
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
weasyprint==59.0
jinja2==3.1.2
rq==1.13.0
redis==4.6.0
pulp==2.7.0
google-api-python-client==2.93.0
google-auth==2.22.0
notion-client==0.3.5
